--- a/benchmark/benchmark_small_payload_test.go
+++ b/benchmark/benchmark_small_payload_test.go
@@ -129,6 +129,17 @@ func BenchmarkJsonParserObjectEachStructSmall(b *testing.B) {
 	}
 }
 
+func BenchmarkJsonParserSetSmall(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		jsonparser.Set(smallFixture, []byte(`"c90927dd-1588-4fe7-a14f-8a8950cfcbd8"`), "uuid")
+		jsonparser.Set(smallFixture, []byte("-3"), "tz")
+		jsonparser.Set(smallFixture, []byte(`"server_agent"`), "ua")
+		jsonparser.Set(smallFixture, []byte("3"), "st")
+
+		nothing()
+	}
+}
+
 /*
    encoding/json
 */
@@ -180,6 +191,19 @@ func BenchmarkGoSimplejsonSmall(b *testing.B) {
 		json.Get("tz").Float64()
 		json.Get("ua").String()
 		json.Get("st").Float64()
+
+		nothing()
+	}
+}
+
+func BenchmarkGoSimplejsonSetSmall(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		json, _ := simplejson.NewJson(smallFixture)
+
+		json.SetPath([]string{"uuid"}, "c90927dd-1588-4fe7-a14f-8a8950cfcbd8")
+		json.SetPath([]string{"tz"}, -3)
+		json.SetPath([]string{"ua"}, "server_agent")
+		json.SetPath([]string{"st"}, 3)
 
 		nothing()
 	}