**Description**:
This PR adds support for `Set` method, enabling end users to use jsonparser to generate json efficiently.
_Note: I am taking this over from #98 after discussion with @ryanleary as he does not have cycles to complete it at the moment._
**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                         50000            170465 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                       200000             29789 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4          500000             18119 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4          500000             21547 ns/op             656 B/op         13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4       200000             31427 ns/op             608 B/op         12 allocs/op
BenchmarkJsonParserSmall-4                       3000000              2414 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4          3000000              2058 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4          2000000              3052 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4       3000000              2545 ns/op             240 B/op          8 allocs/op
```
**Benchmark after change**:
```
BenchmarkJsonParserLarge-4                         50000            147591 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                       300000             28021 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4          500000             18563 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4          500000             19804 ns/op             656 B/op         13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4       300000             29924 ns/op             608 B/op         12 allocs/op
BenchmarkJsonParserSmall-4                       3000000              2089 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4          3000000              2169 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4          2000000              3138 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4       3000000              2822 ns/op             240 B/op          8 allocs/op
BenchmarkJsonParserSetSmall-4                    2000000              4086 ns/op             816 B/op          5 allocs/op
```
Also note compare last to comparable `SetPath()` benchmark in `go-simplejson`:
```
BenchmarkGoSimplejsonSetSmall-4           500000             16755 ns/op            2289 B/op         40 allocs/op
```