@JoshKCarroll 2017-05-11
@bruth I added fixes and tests for the edge cases you mentioned in the other PR, and a couple of others (e.g. trailing whitespace). Also added a couple simple benchmarks. Note sure about the empty byte slice on error but I notice it is mentioned specifically in the documentation.

Let me know if anyone sees other issues, has thoughts or requests for other things to add or change. It has the same limitations about not catching malformed JSON as the underlying `Get()` method, and I did tag it as experimental in the README.