@JoshKCarroll 2017-05-25
FYI @buger we are in process of integrating jsonparser with Set() (and https://github.com/buger/jsonparser/pull/103) into our tool kazaam, definitely putting it through its paces and finding it much improved over simplejson we were using before. https://github.com/qntfy/kazaam/pull/35