@JoshKCarroll 2017-06-01
Fixed a bug where `Set()` in existing empty object would produce an extra comma
```
        {
                desc:    "set unknown key (object, empty partial subtree exists)",
                json:    `{"test":{}}`,
                isFound: true,
                path:    []string{"test", "new.field"},
                setData: `{"key":"new object"}`,
                data:    `{"test":{"new.field":{"key":"new object"}}}`,
        },
```